# docs/swagger.yaml
basePath: /v1/api
definitions:
  handlers.article:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  handlers.articleRequestBody:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  handlers.createArticleResponseBody:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  handlers.message:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  description: A RESTful API for a blogging platform
  title: Blogging Platform API
  version: "1.0"
paths:
  /articles/:
    get:
      consumes:
      - application/json
      description: Fetches all blog articles or blog articles that  match the query
        field
      parameters:
      - description: optional search parameter, it is case insensitive
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: All articles
          schema:
            items:
              $ref: '#/definitions/handlers.article'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.message'
      summary: Gets all articles
      tags:
      - Articles
    post:
      consumes:
      - application/json
      description: Create a new article with the provided title and content
      parameters:
      - description: Article data
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/handlers.articleRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Article created
          schema:
            $ref: '#/definitions/handlers.createArticleResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.message'
      summary: Create a new article
      tags:
      - Articles
  /articles/{articleId}:
    delete:
      consumes:
      - application/json
      description: Deletes the article by deleting the article with the specified
        articleId
      parameters:
      - description: The id of the target article
        in: path
        name: articleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Article deleted
          schema:
            $ref: '#/definitions/handlers.message'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/handlers.message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.message'
      summary: Deletes an article that has the specified id
      tags:
      - Articles
    get:
      description: Fetches an article that matches the id specified in the url
      parameters:
      - description: The id of the article
        in: path
        name: articleId
        required: true
        type: string
      responses:
        "200":
          description: fetched article
          schema:
            $ref: '#/definitions/handlers.article'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/handlers.message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.message'
      summary: fetches article  by ID
      tags:
      - Articles
    put:
      consumes:
      - application/json
      description: Updates the article with the title and content values that have
        been specified, by selecting the article with the specified articleId
      parameters:
      - description: The id of the target article
        in: path
        name: articleId
        required: true
        type: string
      - description: This is the request body containing the fields to be used to
          update the article
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.articleRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: Article updated
          schema:
            $ref: '#/definitions/handlers.article'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/handlers.message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.message'
      summary: Updates an article that has the specified id
      tags:
      - Articles
swagger: "2.0"